Reorganize project into src/ and tests/ folders
- Move source files to src/quality/
- Move test files to tests/
- Add __init__.py files for proper package structure
- Update imports to use src.quality prefix
- Configure pytest in pyproject.toml with pythonpath

New structure:
  src/quality/
    __init__.py
    quality.py
    assembler_recycler_loop.py
    pure_recycler_loop.py
  tests/
    __init__.py
    test_quality.py
    test_loops.py
