+++ src/quality/__init__.py
@@ -1,0 +1,21 @@
+"""Factorio quality system simulation package."""
+
+from src.quality.quality import (
+    BEST_PROD_MODULE,
+    BEST_QUAL_MODULE,
+    NUM_TIERS,
+    QualityTier,
+    create_production_matrix,
+    quality_matrix,
+    quality_probability,
+)
+
+__all__ = [
+    "NUM_TIERS",
+    "BEST_PROD_MODULE",
+    "BEST_QUAL_MODULE",
+    "QualityTier",
+    "quality_probability",
+    "quality_matrix",
+    "create_production_matrix",
+]
